refactor: * move converters - code to protocol and protocol to code - to converts folder.
